requirements.txt
coreapi==2.3.3
Django==2.1.4
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
pyyaml==3.13
